Put Routes Into Separate Routes File
Couldn't find a great / obvious way in Javalin Docs / through Google searching to have routes setup be in a separate file in a Javalin project. So I just made a separate file with a applyRoutes function in it that does the work. I think it's clean enough and does it's job.